add integration test
