+++ go.mod
@@ -7,4 +7,7 @@ require (
 	github.com/decred/dcrd/dcrec/secp256k1/v4 v4.4.0
 )
 
-require github.com/btcsuite/btcd/chaincfg/chainhash v1.0.1 // indirect
+require (
+	github.com/btcsuite/btcd/chaincfg/chainhash v1.0.1 // indirect
+	github.com/joho/godotenv v1.5.1 // indirect
+)
