+++ pkg/account/account.go
@@ -84,7 +84,7 @@ func (a *CEPAccount) UpdateAccount() (bool, error) {
 	}
 
 	// Construct the full URL for the API endpoint
-	url := a.NAGURL + "Circular_GetWalletNonce_" + a.NetworkNode
+	url := fmt.Sprintf("%s/Circular_GetWalletNonce_%s", a.NAGURL, a.NetworkNode)
 
 	// Make the HTTP POST request
 	resp, err := http.Post(url, "application/json", bytes.NewBuffer(jsonData))
@@ -362,6 +362,9 @@ func (a *CEPAccount) SubmitCertificate(cert *Certificate) (map[string]interface{
 // An error is returned if the NAG_URL is not configured, the network request fails,
 // or the JSON response cannot be parsed.
 func (a *CEPAccount) GetTransactionOutcome(TxID string, timeoutSec int) (map[string]interface{}, error) {
+	if a.NAGURL == "" {
+		return nil, fmt.Errorf("network is not set. Please call SetNetwork() first")
+	}
 	startTime := time.Now()
 	timeout := time.Duration(timeoutSec) * time.Second
