+++ pkg/certificate/certificate_test.go
@@ -7,9 +7,7 @@ import (
 	"fmt"
 	"io"
 	"net/http"
-	"net/http/httptest"
 	"reflect"
-	"strings"
 	"testing"
 )
 
@@ -287,144 +285,4 @@ func (acc *CEPAccount) SubmitCertificate(cert *Certificate) (map[string]interfac
 		return nil, fmt.Errorf("failed to decode response JSON: %w", err)
 	}
 	return result, nil
-}
-
-func TestSubmitCertificate(t *testing.T) {
-	testCases := []struct {
-		name           string
-		nagURL         string
-		mockResponse   string
-		mockStatusCode int
-		cert           *Certificate
-		expectError    bool
-		expectedError  string
-		expectedResult map[string]interface{}
-	}{
-		{
-			name:           "Successful Certificate Submission",
-			nagURL:         "http://localhost:8080",
-			mockResponse:   `{"txHash": "0xabc123", "status": "success"}`,
-			mockStatusCode: http.StatusOK,
-			cert: &Certificate{
-				Version: "1.0",
-				Data:    "test data",
-			},
-			expectError:    false,
-			expectedResult: map[string]interface{}{"txHash": "0xabc123", "status": "success"},
-		},
-		{
-			name:           "NAGURL Not Set",
-			nagURL:         "",
-			mockResponse:   "",
-			mockStatusCode: http.StatusOK,
-			cert: &Certificate{
-				Version: "1.0",
-				Data:    "test data",
-			},
-			expectError:   true,
-			expectedError: "network is not set. Please call SetNetwork() first",
-		},
-		{
-			name:           "HTTP Error - Bad Status",
-			nagURL:         "http://localhost:8080",
-			mockResponse:   "Internal Server Error",
-			mockStatusCode: http.StatusInternalServerError,
-			cert: &Certificate{
-				Version: "1.0",
-				Data:    "test data",
-			},
-			expectError:   true,
-			expectedError: "network returned an error - status: 500 Internal Server Error, body: Internal Server Error",
-		},
-		{
-			name:           "Invalid JSON Response",
-			nagURL:         "http://localhost:8080",
-			mockResponse:   `{"txHash": "0xabc123", "status": "success"`, // Malformed JSON
-			mockStatusCode: http.StatusOK,
-			cert: &Certificate{
-				Version: "1.0",
-				Data:    "test data",
-			},
-			expectError:   true,
-			expectedError: "failed to decode response JSON",
-		},
-		{
-			name:           "Empty Certificate Data",
-			nagURL:         "http://localhost:8080",
-			mockResponse:   `{"txHash": "0xabc123", "status": "success"}`,
-			mockStatusCode: http.StatusOK,
-			cert:           &Certificate{}, // Empty certificate
-			expectError:    false,
-			expectedResult: map[string]interface{}{"txHash": "0xabc123", "status": "success"},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			// Create a mock HTTP server if a NAGURL is provided
-			var server *httptest.Server
-			var capturedRequestBody []byte
-			if tc.nagURL != "" {
-				server = httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-					var err error
-					capturedRequestBody, err = io.ReadAll(r.Body)
-					if err != nil {
-						t.Fatalf("Failed to read request body: %v", err)
-					}
-					w.WriteHeader(tc.mockStatusCode)
-					w.Write([]byte(tc.mockResponse))
-				}))
-				defer server.Close()
-			}
-
-			// Import the necessary constants and NewCEPAccount from the appropriate package if available.
-			// For example:
-			// import "Circular-Protocol/CEP-Go-APIs/pkg/circular_enterprise_apis"
-			// acc := circular_enterprise_apis.NewCEPAccount(circular_enterprise_apis.DefaultNAG, circular_enterprise_apis.DefaultChain, circular_enterprise_apis.LibVersion)
-
-			// If the above import is set up, use:
-			// acc := circular_enterprise_apis.NewCEPAccount(circular_enterprise_apis.DefaultNAG, circular_enterprise_apis.DefaultChain, circular_enterprise_apis.LibVersion)
-
-			// Otherwise, fallback to the local definition (as before):
-			acc := NewCEPAccount(DefaultNAG, DefaultChain, LibVersion)
-			if tc.nagURL != "" {
-				acc.NAGURL = server.URL
-			} else {
-				acc.NAGURL = tc.nagURL
-			}
-
-			result, err := acc.SubmitCertificate(tc.cert)
-
-			if tc.expectError {
-				if err == nil {
-					t.Errorf("Expected an error but got nil")
-				}
-				if !strings.Contains(err.Error(), tc.expectedError) {
-					t.Errorf("Expected error message to contain '%s', but got '%s'", tc.expectedError, err.Error())
-				}
-			} else {
-				if err != nil {
-					t.Errorf("Expected no error, but got: %v", err)
-				}
-				// Compare the result map
-				if len(result) != len(tc.expectedResult) {
-					t.Errorf("Expected result map length %d, but got %d", len(tc.expectedResult), len(result))
-				}
-				for k, v := range tc.expectedResult {
-					if result[k] != v {
-						t.Errorf("Expected result[%s] to be %v, but got %v", k, v, result[k])
-					}
-				}
-
-				// Verify the request body sent to the mock server
-				expectedRequestBody, marshalErr := json.Marshal(tc.cert)
-				if marshalErr != nil {
-					t.Fatalf("Failed to marshal expected certificate: %v", marshalErr)
-				}
-				if !bytes.Equal(capturedRequestBody, expectedRequestBody) {
-					t.Errorf("Request body mismatch. Expected %s, got %s", string(expectedRequestBody), string(capturedRequestBody))
-				}
-			}
-		})
-	}
 }
